Create module logAndStartTrackingBalances to make testing easier
