+++ src/run/logAndStartTrackingBalances.ts
@@ -1,3 +1,5 @@
+import { formatEther } from "@ethersproject/units";
+import { gray, yellow } from "chalk";
 import {
   Contract,
   ContractInterface,
@@ -6,6 +8,8 @@ import {
   Signer,
 } from "ethers";
 import snx from "synthetix";
+import { trackKeeperBalance } from "../metrics";
+
 const { getSource, getTarget } = snx;
 
 // This is lifted from the synthetix-js package, since the package doesn't
@@ -55,5 +59,56 @@ const getSynthetixContracts = ({
       return acc;
     }, {});
 };
+async function logAndStartTrackingBalances(
+  {
+    network,
+    provider,
+    signers,
+  }: {
+    network: string;
+    provider: providers.WebSocketProvider | providers.JsonRpcProvider;
+    signers: Signer[];
+  },
+  deps = { getSynthetixContracts, trackKeeperBalance }
+) {
+  // Check balances of accounts.
+  const { SynthsUSD } = deps.getSynthetixContracts({
+    network,
+    provider: provider,
+    useOvm: true,
+  });
 
-export default getSynthetixContracts;
+  const signerBalances = await Promise.all(
+    signers.map(async signer => {
+      // ETH.
+      const balance = await signer.getBalance();
+
+      const address = await signer.getAddress();
+      // sUSD.
+      const sUSDBalance = await SynthsUSD.balanceOf(address);
+
+      const balances = [
+        ["ETH", balance],
+        ["sUSD", sUSDBalance],
+      ];
+
+      return { balances, address };
+    })
+  );
+  // Log and track account balances
+  signerBalances.forEach(({ address, balances }, i) => {
+    const balanceText = balances
+      .map(([key, balance]) => {
+        let balanceText = formatEther(balance);
+        if (balance.isZero()) {
+          balanceText = yellow(balanceText);
+        }
+        return `${balanceText} ${key}`;
+      })
+      .join(", ");
+
+    console.log(gray(`Account #${i}: ${address} (${balanceText})`));
+    deps.trackKeeperBalance(signers[i], SynthsUSD);
+  });
+}
+export default logAndStartTrackingBalances;
